Cross-sections

Wires.CrossSection — Type

CrossSection

Abstract type for wire cross-sections.

Wires.Circular — Type

Circular{T}

A circular wire cross-section.

Fields

diameter::T: The diameter of the wire.

Wires.Rectangular — Type

Rectangular{T}

A rectangular wire cross-section.

Fields

width::T: The width of the wire.

height::T: The height of the wire.

Wires.RectangularHollow — Function

RectangularHollow(; kwargs...) -> Difference{Rectangular, Circular}

Returns the difference of a rectangular wire cross-section and a circular wire cross-section with the given width, height, and hole diameter.

Keywords

width::Unitful.Length: The width of the wire.

height::Unitful.Length: The height of the wire.

hole_diameter::Unitful.Length: The diameter of the hole.

Returns

difference::Difference{Rectangular, Circular}: The difference representing a rectangular hollow wire cross-section.

Wires.area — Function

area(s::CrossSection)

Returns the area of the given wire cross-section.

Arguments

s::CrossSection: The wire cross-section.

Returns

Unitful.Area: The area of the wire cross-section.

Wires.resistance — Function

resistance(s::CrossSection; m::Material = Copper)

Returns the resistance of the given wire cross-section assuming copper as the material by default.

Arguments

s::CrossSection: The wire cross-section.

Keywords

m::Material = Copper: The material of the wire.

Returns

Unitful.ElectricalResistance: The resistance of the wire cross-section.

Wires.heat_capacity — Function

heat_capacity(s::CrossSection; m::Material = Copper)

Returns the heat capacity of the given wire cross-section assuming copper as the material by default.

Arguments

s::CrossSection: The wire cross-section.

Keywords

m::Material = Copper: The material of the wire.

Returns

Unitful.SpecificHeatCapacity: The heat capacity of the wire cross-section.