Materials

Wires.Material — Type

Material{T1,T2,T3}

A material with a name, resistivity, density, and heat capacity.

Fields

name::String: The name of the material.

resistivity::T1: The resistivity of the material.

density::T2: The density of the material.

heat_capacity::T3: The heat capacity of the material.

Wires.Copper — Constant

Copper

Copper as instance of Material with properties from [Wikipedia][1].

[1]: https://en.wikipedia.org/wiki/Electricalresistivityandconductivity#Resistivityandconductivityofvariousmaterials